+++ dflib/src/main/java/org/dflib/LongSeries.java
@@ -44,6 +44,23 @@ public interface LongSeries extends Series<Long> {
     LongSeries materialize();
 
     @Override
+    default int position(Long value) {
+        if (value == null) {
+            return -1;
+        }
+
+        long pval = value;
+        int len = size();
+        for (int i = 0; i < len; i++) {
+            if (pval == getLong(i)) {
+                return i;
+            }
+        }
+
+        return -1;
+    }
+
+    @Override
     default Series<?> expand(Object... values) {
         int len = values.length;
         if (len == 0) {