+++ RELEASE-NOTES.md
@@ -8,6 +8,7 @@
 * #270 Replace custom predicates with the ones from JDK
 * #271 Make Index more Series-like
 * #272 GroupBy.select(Exp...) / map(Exp...)
+* #273 Series.replace(Map), position(..), contains(..), expand(..)
 
 ## Release 1.0.0-M20